#35 [REFACTOR] review 폴더 구조 개선
